+++ open-coroutine-timer/src/lib.rs
@@ -46,6 +46,8 @@
     clippy::single_char_lifetime_names, // TODO: change lifetime names
 )]
 
+use std::collections::vec_deque::{Iter, IterMut};
+use std::collections::VecDeque;
 use std::time::{Duration, SystemTime, UNIX_EPOCH};
 
 const NANOS_PER_SEC: u64 = 1_000_000_000;
@@ -85,16 +87,155 @@ pub fn add_timeout_time(time: u64) -> u64 {
     }
 }
 
-mod typed;
+#[derive(Debug, PartialEq, Eq)]
+pub struct TimerEntry<T> {
+    time: u64,
+    inner: VecDeque<T>,
+}
 
-pub use typed::*;
+impl<T> TimerEntry<T> {
+    #[must_use]
+    pub fn new(time: u64) -> Self {
+        TimerEntry {
+            time,
+            inner: VecDeque::new(),
+        }
+    }
+
+    #[must_use]
+    pub fn len(&self) -> usize {
+        self.inner.len()
+    }
+
+    #[must_use]
+    pub fn is_empty(&self) -> bool {
+        self.inner.is_empty()
+    }
+
+    #[must_use]
+    pub fn get_time(&self) -> u64 {
+        self.time
+    }
+
+    pub fn pop_front(&mut self) -> Option<T> {
+        self.inner.pop_front()
+    }
+
+    pub fn push_back(&mut self, t: T) {
+        self.inner.push_back(t);
+    }
+
+    pub fn remove(&mut self, t: &T) -> Option<T>
+    where
+        T: Ord,
+    {
+        let index = self
+            .inner
+            .binary_search_by(|x| x.cmp(t))
+            .unwrap_or_else(|x| x);
+        self.inner.remove(index)
+    }
+
+    pub fn iter_mut(&mut self) -> IterMut<'_, T> {
+        self.inner.iter_mut()
+    }
+
+    #[must_use]
+    pub fn iter(&self) -> Iter<'_, T> {
+        self.inner.iter()
+    }
+}
+
+#[repr(C)]
+#[derive(Debug, PartialEq, Eq)]
+pub struct TimerList<T> {
+    dequeue: VecDeque<TimerEntry<T>>,
+}
+
+impl<T> TimerList<T> {
+    #[must_use]
+    pub fn new() -> Self {
+        TimerList {
+            dequeue: VecDeque::new(),
+        }
+    }
+
+    #[must_use]
+    pub fn len(&self) -> usize {
+        self.dequeue.len()
+    }
+
+    pub fn insert(&mut self, time: u64, t: T) {
+        let index = self
+            .dequeue
+            .binary_search_by(|x| x.time.cmp(&time))
+            .unwrap_or_else(|x| x);
+        if let Some(entry) = self.dequeue.get_mut(index) {
+            entry.push_back(t);
+        } else {
+            let mut entry = TimerEntry::new(time);
+            entry.push_back(t);
+            self.dequeue.insert(index, entry);
+        }
+    }
+
+    #[must_use]
+    pub fn front(&self) -> Option<&TimerEntry<T>> {
+        self.dequeue.front()
+    }
+
+    pub fn pop_front(&mut self) -> Option<TimerEntry<T>> {
+        self.dequeue.pop_front()
+    }
+
+    #[must_use]
+    pub fn is_empty(&self) -> bool {
+        self.dequeue.is_empty()
+    }
+
+    pub fn get_entry(&mut self, time: u64) -> Option<&mut TimerEntry<T>> {
+        let index = self
+            .dequeue
+            .binary_search_by(|x| x.time.cmp(&time))
+            .unwrap_or_else(|x| x);
+        self.dequeue.get_mut(index)
+    }
+
+    pub fn iter_mut(&mut self) -> IterMut<'_, TimerEntry<T>> {
+        self.dequeue.iter_mut()
+    }
+
+    #[must_use]
+    pub fn iter(&self) -> Iter<'_, TimerEntry<T>> {
+        self.dequeue.iter()
+    }
+}
+
+impl<T> Default for TimerList<T> {
+    fn default() -> Self {
+        Self::new()
+    }
+}
 
 #[cfg(test)]
 mod tests {
-    use crate::now;
+    use super::*;
 
     #[test]
     fn test() {
         println!("{}", now());
+    }
+
+    #[test]
+    fn timer_list() {
+        let mut list = TimerList::new();
+        assert_eq!(list.len(), 0);
+        list.insert(1, String::from("data is typed"));
+        assert_eq!(list.len(), 1);
+
+        let mut entry = list.pop_front().unwrap();
+        assert_eq!(entry.len(), 1);
+        let string = entry.pop_front().unwrap();
+        assert_eq!(string, String::from("data is typed"));
     }
 }
